use the preprocessr better
