+++ fabric/mainProject
@@ -1,1 +1,1 @@
-1.21.11
+26.1.2
